Fix plottable program test
